requirements.txt
Flask==2.3.2
google-api-core==2.11.0
google-api-python-client==2.95.0
google-auth==2.17.3
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.59.0
MarkupSafe==2.1.3
gunicorn==21.2.0